+++ src/V1/errors/InvalidUuid.d.ts
@@ -1,6 +1,0 @@
-export declare class InvalidUuidError {
-    readonly invalidInput: string;
-    constructor(invalidInput: string);
-}
-export declare function IsInvalidUuidError(input: any): input is InvalidUuidError;
-//# sourceMappingURL=InvalidUuid.d.ts.map