+++ src/V1/errors/index.d.ts
@@ -1,2 +1,0 @@
-export * from './InvalidUuid';
-//# sourceMappingURL=index.d.ts.map